requirements.txt
asgiref==3.11.1
Django==5.2.13
djangorestframework==3.17.1
python-decouple==3.8
sqlparse==0.5.5
tzdata==2026.1
